+++ src/entities/DronerEntities.ts
@@ -17,6 +17,7 @@ export interface DronerEntity {
   dronerCode: string;
   firstname: string;
   lastname: string;
+  nickname: string;
   idNo: string;
   telephoneNo: string;
   status: string;
@@ -41,12 +42,14 @@ export interface DronerEntity {
   comment?: string;
   createBy?: string;
   dateWaitPending: string,
+  isDelete: boolean;
 }
 export const DronerEntity_INIT: DronerEntity = {
   id: "",
   dronerCode: "",
   firstname: "",
   lastname: "",
+  nickname: "",
   idNo: "",
   telephoneNo: "",
   status: "",
@@ -68,7 +71,8 @@ export const DronerEntity_INIT: DronerEntity = {
   updatedAt: "",
   totalDroneCount: 0,
   birthDate: "",
-  dateWaitPending: ""
+  dateWaitPending: "",
+  isDelete: false
 };
 export interface CreateDronerEntity {
   id?: string;