+++ src/entities/FarmerEntities.ts
@@ -20,6 +20,7 @@ export interface FarmerEntity {
   id: string;
   firstname: string;
   lastname: string;
+  nickname: string;
   idNo: string;
   telephoneNo: string;
   status: string;
@@ -37,6 +38,7 @@ export const FarmerEntity_INIT: FarmerEntity = {
   id: "",
   firstname: "",
   lastname: "",
+  nickname: "",
   idNo: "",
   telephoneNo: "",
   status: "",