+++ src/pages/promotion/EditPromotionPage.tsx
@@ -69,6 +69,7 @@ function EditPromotion() {
   const [plantName, setPlantName] = useState<CropPurposeSprayEntity[]>();
   const [provinceList, setProvinceList] = useState<string[]>([]);
   const [countSelectFarmer, setCountSelectFarmer] = useState<any>();
+  const [editPromo , setEditPromo] = useState<boolean>(false);
   const [crop, setCrop] = useState<any>([
     {
       id: null,
@@ -80,7 +81,7 @@ function EditPromotion() {
   const [descriptionEditor, setDescriptionEditor] = useState<string | null>(
     null
   );
-  const [checkDelete, setCheckDelete] = useState<any>();
+  const [defaultCount, setDefaultCount] = useState<any>();
   const [conditionEditor, setConditionEditor] = useState<string | null>(null);
   const [editTable, setEditTable] = useState(true);
   const [province, setProvince] = useState<string[]>([]);
@@ -200,7 +201,6 @@ function EditPromotion() {
       setRow(row + 10);
     }
   };
-
   const getPromotion = (id: string) => {
     CouponDataSource.queryCoupon(id)
       .then(async (res) => {
@@ -238,6 +238,8 @@ function EditPromotion() {
           provinceCheckbox: res.couponConditionProvince,
           couponConditionProvinceList: res.couponConditionProvinceList,
         });
+        setDefaultCount(res.count);
+        setEditPromo(res.keep > 0);
         setConditionSpecialFirsttime(res.conditionSpecialFirsttime ?? false);
         setRaiCondition(res.couponConditionRai ?? false);
         setServiceCondition(res.couponConditionService ?? false);
@@ -681,7 +683,7 @@ function EditPromotion() {
     if (renderMobile.expiredDate) {
       const getValueDate = form.getFieldsValue();
       const endDate = moment(getValueDate.DateExpired).format("YYYY-MM-DD");
-      return current && current > dayjs(endDate);
+      return current && current < dayjs(endDate);
     }
   };
   const handleDescriptionEditor = (
@@ -1100,6 +1102,7 @@ function EditPromotion() {
                     ]}
                   >
                     <AntdSelect
+                      disabled={editPromo}
                       className="col-lg-12 p-1"
                       placeholder="เลือกประเภทคูปอง"
                       onChange={handleCouponType}
@@ -1136,7 +1139,7 @@ function EditPromotion() {
                           },
                         ]}
                       >
-                        <Radio.Group className="d-flex" onChange={handleCoupon}>
+                        <Radio.Group className="d-flex" onChange={handleCoupon} disabled={editPromo}>
                           <Radio value={"FREE"}>ฟรีค่าบริการ</Radio>
                           <Radio value={"DISCOUNT"}>ส่วนลด</Radio>
                         </Radio.Group>
@@ -1153,7 +1156,7 @@ function EditPromotion() {
                         ]}
                       >
                         <Input
-                          disabled={coupon !== "DISCOUNT"}
+                          disabled={coupon !== "DISCOUNT" || editPromo}
                           placeholder="กรอกจำนวนเงิน"
                           autoComplete="off"
                           onChange={(e) => checkNumber(e, "discount")}
@@ -1202,7 +1205,7 @@ function EditPromotion() {
                 </div>
                 <div className="form-group col-lg-6">
                   <label>
-                    จำนวนสิทธิ์ <span style={{ color: "red" }}>*</span>
+                    จำนวนสิทธิ์ <span style={{ color: "red" }}>* (สามารถปรับเพิ่มได้เท่านั้น)</span>
                   </label>
                   <div className="mt-1">
                     {couponConditionFarmerList.length > 0 ? (
@@ -1264,6 +1267,7 @@ function EditPromotion() {
                         }}
                         format={dateFormat}
                         disabledDate={disabledStartDate}
+                        disabled={editPromo}
                       />
                     </Form.Item>
                     <Form.Item
@@ -1276,6 +1280,7 @@ function EditPromotion() {
                       ]}
                     >
                       <TimePicker
+                      disabled={editPromo}
                         format={"HH:mm"}
                         className="ms-3"
                         placeholder="เลือกเวลา"
@@ -1294,7 +1299,7 @@ function EditPromotion() {
                 </div>
                 <div className="form-group col-lg-6">
                   <label>
-                    วันสิ้นสุด <span style={{ color: "red" }}>*</span>
+                    วันสิ้นสุด <span style={{ color: "red" }}>* (สามารถขยายเวลาเพิ่มได้เท่านั้น)</span>
                   </label>
                   <div className="d-flex">
                     <Form.Item
@@ -1412,6 +1417,7 @@ function EditPromotion() {
                     valuePropName="checked"
                   >
                     <Checkbox
+                    disabled={editPromo}
                       onChange={handleConditionSpecialFirsttime}
                       checked={conditionSpecialFirsttime}
                       className="pt-3"
@@ -1426,7 +1432,7 @@ function EditPromotion() {
                           onChange={handleSpecificFarmer}
                           checked={specificFarmer}
                           className="pt-3"
-                          disabled={couponConditionFarmerList.length > 0}
+                          disabled={couponConditionFarmerList.length > 0 || editPromo}
                         >
                           ให้เฉพาะเกษตรกรบางคน
                         </Checkbox>
@@ -1491,6 +1497,7 @@ function EditPromotion() {
                   <div>
                     <Form.Item name="raiCheckbox" valuePropName="checked">
                       <Checkbox
+                      disabled={editPromo}
                         onChange={handleRaiCondition}
                         checked={raiCondition}
                         className="pt-3"
@@ -1500,7 +1507,7 @@ function EditPromotion() {
                             <label>จำนวนไร่ขั้นต่ำ</label>
                             <Form.Item name="couponConditionRaiMin">
                               <Input
-                                disabled={!raiCondition}
+                                disabled={!raiCondition || editPromo}
                                 placeholder="กรอกจำนวนไร่"
                                 onChange={(e) => {
                                   checkNumber(e, "couponConditionRaiMin");
@@ -1516,7 +1523,7 @@ function EditPromotion() {
                             <label>จำนวนไร่สูงสุด</label>
                             <Form.Item name="couponConditionRaiMax">
                               <Input
-                                disabled={!raiCondition}
+                                disabled={!raiCondition || editPromo}
                                 placeholder="กรอกจำนวนไร่"
                                 onChange={(e) => {
                                   checkNumber(e, "couponConditionRaiMax");
@@ -1535,6 +1542,7 @@ function EditPromotion() {
                   <div>
                     <Form.Item name="serviceCheckbox" valuePropName="checked">
                       <Checkbox
+                      disabled={editPromo}
                         onChange={handleServiceCondition}
                         checked={serviceCondition}
                         className="pt-3"
@@ -1544,7 +1552,7 @@ function EditPromotion() {
                             <label>จำนวนค่าบริการขั้นต่ำ</label>
                             <Form.Item name="couponConditionServiceMin">
                               <Input
-                                disabled={!serviceCondition}
+                                disabled={!serviceCondition || editPromo}
                                 placeholder="กรอกค่าบริการ"
                                 onChange={(e) => {
                                   checkNumber(e, "couponConditionServiceMin");
@@ -1560,7 +1568,7 @@ function EditPromotion() {
                             <label>จำนวนค่าบริการสูงสุด</label>
                             <Form.Item name="couponConditionServiceMax">
                               <Input
-                                disabled={!serviceCondition}
+                                disabled={!serviceCondition || editPromo}
                                 placeholder="กรอกค่าบริการ"
                                 onChange={(e) => {
                                   checkNumber(e, "couponConditionServiceMax");
@@ -1582,6 +1590,7 @@ function EditPromotion() {
                 <div className="form-group col-lg-12 d-flex justify-content-between align-items-center pb-4">
                   <Form.Item name="plantCheckbox" valuePropName="checked">
                     <Checkbox
+                    disabled={editPromo}
                       className="pt-3"
                       checked={couponPlant}
                       onChange={handlePlantCoupon}
@@ -1606,6 +1615,7 @@ function EditPromotion() {
                 <div className="form-group col-lg-12 d-flex justify-content-between align-items-center pt-2 pb-3">
                   <Form.Item name="provinceCheckbox" valuePropName="checked">
                     <Checkbox
+                    disabled={editPromo}
                       checked={couponProvince}
                       onChange={handleCouponProvince}
                       className="pt-3"
@@ -1634,7 +1644,7 @@ function EditPromotion() {
                     ]}
                   >
                     <AntdSelect
-                      disabled={!couponProvince}
+                      disabled={!couponProvince|| editPromo}
                       mode="multiple"
                       placeholder="เลือกจังหวัด"
                       onChange={handleChangeProvince}