+++ src/pages/task/newTask/AddNewTask.tsx
@@ -91,6 +91,7 @@ import type { GroupBase, OptionsOrGroups } from "react-select";
 import { FarmerPageEntity } from "../../../entities/FarmerEntities";
 import { InputPicker, SelectPicker } from "rsuite";
 import "rsuite/dist/rsuite.min.css";
+import ShowNickName from "../../../components/popover/ShowNickName";
 export type OptionType = {
   value: any;
   label: any;
@@ -1325,13 +1326,12 @@ const AddNewTask = () => {
   const handleVisibleRating = (newVisible: any) => {
     setVisibleRating(newVisible);
   };
-
   const searchSection = (
     <div className="d-flex justify-content-between" style={{ padding: "10px" }}>
       <div className="col-lg-3 p-1">
         <Input
           prefix={<SearchOutlined style={{ color: color.Disable }} />}
-          placeholder="ค้นหาชื่อเกษตรกร หรือเบอร์โทร"
+          placeholder="ค้นหาชื่อนักบินโดรน หรือเบอร์โทร"
           className="col-lg-12"
           onChange={(e: any) => setSearchTextDroner(e.target.value)}
         />
@@ -1503,7 +1503,12 @@ const AddNewTask = () => {
                 </Tooltip>
               )}
               <br />
-              <span style={{ color: color.Grey }}>{row.droner_code}</span>
+              <span style={{ color: color.Grey }}>
+                {row.droner_code}
+                {row.nickname && (
+                  <ShowNickName data={row.nickname} menu="INFO" />
+                )}
+              </span>
             </>
           ),
         };
@@ -1711,6 +1716,12 @@ const AddNewTask = () => {
                         }}
                       >
                         {JSON.parse(x).droner_code}
+                        {JSON.parse(x).nickname && (
+                          <ShowNickName
+                            data={JSON.parse(x).nickname}
+                            menu="INFO"
+                          />
+                        )}
                       </p>
                     </div>
                     <div className="col-lg-2">{JSON.parse(x).telephone_no}</div>